requirements.txt
aiobotocore==2.15.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aioitertools==0.12.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
av==13.1.0
beautifulsoup4==4.12.3
beir==2.0.0
blis==0.7.11
boto3==1.35.23
botocore==1.35.23
braceexpand==0.1.7
cachetools==5.5.0
catalogue==2.0.10
cbor==1.0.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
datasets==2.21.0
deepspeed==0.15.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.0
elasticsearch==7.9.1
# en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
eval_type_backport==0.2.0
evaluate==0.4.3
exceptiongroup==1.2.2
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec>=2024.3.1
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
hjson==3.1.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
ijson==3.3.0
iniconfig==2.0.0
inscriptis==2.5.0
Jinja2==3.1.4
jiter==0.6.0
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
lightgbm==4.5.0
littleutils==0.2.4
loguru==0.7.2
lxml==5.3.0
lz4==4.3.3
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
mteb==1.8.11
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.10
networkx==3.3
ninja==1.11.1.1
nltk==3.9.1
nomic==3.3.3
numpy==1.26.4
ogb==1.3.6
onnx==1.17.0
onnxconverter-common==1.14.0
onnxruntime==1.19.2
openai==1.51.1
outdated==0.2.2
packaging==24.1
pandas==2.2.3
pathlib_abc==0.1.1
pathy==0.11.0
peft==0.13.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
polars==1.9.0
preshed==3.0.9
pretty-errors==1.2.25
proto-plus==1.24.0
protobuf==3.20.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycocoevalcap==1.2
pycocotools==2.0.8
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyjnius==1.6.1
PyJWT==2.9.0
pynvml==11.5.3
pyparsing==3.1.4
pytest==8.3.3
python-dateutil==2.9.0.post0
pytrec-eval==0.5
pytrec-eval-terrier==0.5.6
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.2
rsa==4.9
s3fs==2024.6.1
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==3.3.0
sentencepiece==0.2.0
sentry-sdk==2.15.0
setproctitle==1.3.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.3
tabulate==0.9.0
thinc==8.1.12
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.9
tokenizers==0.20.0
tomli==2.0.2
torch==2.4.1
torchaudio==2.4.1
torchmetrics==1.4.2
tqdm==4.66.5
transformers==4.45.2
triton==3.0.0
typer==0.9.4
typing_extensions==4.12.2
tzdata==2024.2
unlzw3==0.2.3
urllib3==2.2.3
wandb==0.18.3
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wasabi==1.1.3
wcwidth==0.2.13
webdataset==0.2.100
wilds==2.0.0
wrapt==1.16.0
xxhash==3.5.0
yarl==1.13.1
zlib-state==0.1.9
